requirements.txt
igraph
matplotlib
numpy
cupy-cuda12x
